flaskbackend/requirements.txt
# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.1.0
python-multipart==0.0.17
python-dotenv==1.0.1

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography>=42.0.0
bcrypt>=4.1.0
argon2-cffi>=23.1.0
email-validator==2.1.0

# ============================================================================
# DATABASE
# ============================================================================
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# ============================================================================
# SUPABASE INTEGRATION
# ============================================================================
supabase>=2.0.0
postgrest>=0.13.0

# ============================================================================
# API & HTTP
# ============================================================================
requests>=2.31.0
urllib3>=2.1.0
httpx>=0.25.0
aiohttp>=3.9.0
fastapi-cache2==0.2.1

# ============================================================================
# PAYMENT GATEWAYS
# ============================================================================
stripe>=5.0.0
paypalrestsdk>=1.13.1
pycryptodome>=3.18.0

# ============================================================================
# DATA SCIENCE & MACHINE LEARNING
# ============================================================================
numpy>=1.24.0,<2.0
pandas==2.1.2
scikit-learn==1.5.2
scipy==1.11.3
scikit-image==0.22.0
statsmodels==0.14.0
joblib==1.4.2

# ============================================================================
# DEEP LEARNING (Optional - comment out if not using GPU)
# ============================================================================
# tensorflow>=2.16.1,<2.21
# keras>=3.0.0
# tensorboard>=2.16.0

# PyTorch (CPU version - for GPU, see PyTorch website)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ============================================================================
# COMPUTER VISION
# ============================================================================
opencv-python-headless>=4.8.0  # Headless version (no GUI, smaller)
pillow>=10.0.0

# YOLOv8 and Object Detection
ultralytics>=8.0.0
albumentations>=1.3.0
timm>=0.9.0

# ============================================================================
# MODEL EXPORT & OPTIMIZATION
# ============================================================================
onnx>=1.15.0
onnxruntime>=1.16.0
h5py>=3.10.0

# Mobile deployment (TensorFlow Lite, CoreML)
# Note: TensorFlow is optional - uncomment only if needed for TFLite export
# tensorflow>=2.16.1  # For TFLite conversion (large package ~500MB)
# tf2onnx>=1.16.0  # Alternative: TensorFlow to ONNX converter

# CoreML export for iOS (macOS only)
# coremltools>=7.0  # Uncomment on macOS for iOS model export

# Model optimization
imagehash>=4.3.1  # Perceptual image hashing for duplicate detection

# ============================================================================
# NATURAL LANGUAGE PROCESSING
# ============================================================================
nltk>=3.8.0

# ============================================================================
# MACHINE LEARNING & AUTOML
# ============================================================================
optuna>=3.5.0
ray[tune]>=2.9.0

# ============================================================================
# VISUALIZATION
# ============================================================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ============================================================================
# UTILITIES
# ============================================================================
networkx>=3.2.0
pytz>=2023.3
python-dateutil>=2.8.2

# ============================================================================
# MONITORING & LOGGING
# ============================================================================
tenacity==8.2.3
prometheus-client==0.17.1
sentry-sdk==1.32.0

# ============================================================================
# BACKGROUND TASKS
# ============================================================================
celery==5.3.4
flower==2.0.1

# ============================================================================
# EMAIL
# ============================================================================
emails==0.6.0

# ============================================================================
# TESTING (Optional - uncomment if needed)
# ============================================================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0

# ============================================================================
# DEVELOPMENT TOOLS (Optional - uncomment if needed)
# ============================================================================
# black>=23.12.0
# flake8>=6.1.0
# mypy>=1.7.0
# ipython>=8.18.0
# jupyter>=1.0.0